Questions 10, 14 and 16
Question 10

Played around with object types, ids and methods. Haven't really done
anything too relevant to answering the question.

Question 14

Wrote the JavaScript code for the primeFinder function.
Will need to go back and make some changes, such as only checking for a
divisor that is half the size of the number being divided.

Question 16
Just messing around a little but more, mostly with the way that it
looks.